Work on the AST and overall compilation/optimization/validation processes.

+++ src/vc/basic/sourcelocation.h
@@ -1,11 +1,13 @@
 #ifndef VC_SOURCE_LOCATION_H_INCLUDE
 #define VC_SOURCE_LOCATION_H_INCLUDE
 
+#include <vc/types.h>
+
 namespace vc {
 
 struct SourceLocation {
-    int line;
-    int column;
+    u64 line;
+    u64 column;
 };
 
 struct SourceRange {
